requirements.txt
discord.py~=2.6.4
flask~=3.0.3